tour_firm_project/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
